openGLPython/requirements.txt
PyOpenGL 
PyOpenGL_accelerate
glfw
numpy
argparse
opencv-python